requirements.txt
requests
beautifulsoup4
lxml
trafilatura
python-docx
python-dotenv
